Extract command palette filter to a non-component module (eslint react-refresh) (W1-β review)
